deployment/cdk/requirements.txt
# AWS CDK Core
aws-cdk-lib>=2.100.0,<3.0.0
constructs>=10.0.0,<11.0.0

# Additional dependencies for CDK stacks
boto3>=1.28.0
